+++ spec/services/script_generator_spec.rb
@@ -53,6 +53,12 @@ describe ScriptGenerator do
 
       expect { generator.call }.to raise_error(ArgumentError)
     end
+
+    it 'inserts case number of the patient' do
+      script = ScriptGenerator.new(computation, '<%= patient.case_number %>').call
+
+      expect(script).to include(patient.case_number)
+    end
   end
 
   it 'inserts download file curl' do